requirements.txt
openai==1.37.1
aiogram==2.25.1
aiocron==1.8
python-dotenv==1.0.0